+++ FilmBegetter.WEB/ClientApp/src/app/core/models/authenticationResponseViewModel.interface.ts
