+++ FilmBegetter.WEB/ClientApp/src/app/core/models/authenticationViewModel.interface.ts
