+++ FilmBegetter.WEB/ClientApp/src/app/core/models/genreViewModel.interface.ts
