+++ FilmBegetter.WEB/ClientApp/src/app/core/models/movieViewModel.interface.ts
