+++ FilmBegetter.WEB/ClientApp/src/app/core/models/registrationResponseViewModel.interface.ts
