+++ FilmBegetter.WEB/ClientApp/src/app/core/models/registrationViewModel.interface.ts
