+++ FilmBegetter.WEB/ClientApp/src/app/shared/components/card/component/card.component.ts
@@ -2,7 +2,6 @@ import { IUserCard } from './../../../models/card-user.interface';
 import { IStatistics } from './../../../models/card-statistics.interface copy';
 import { IMovieCard } from './../../../models/card.interface';
 import { Component, Input, OnInit } from '@angular/core';
-import {MovieViewModel} from "../../../models/movieViewModel.interface";
 
 @Component({
     selector: 'app-card',
