+++ FilmBegetter.WEB/ClientApp/src/app/shared/models/card.interface.ts
@@ -1,5 +1,4 @@
-import { IMovie } from './../../core/models/movie.interface';
-import { MovieViewModel } from "./movieViewModel.interface";
+import {MovieViewModel} from "../../core/models/movieViewModel.interface";
 
 export interface IMovieCard {
     type: 'defaultPreview' | 'smallPreview' | 'adminView',
